+++ packages/verified-fetch/CHANGELOG.md
@@ -1,3 +1,50 @@
+## @helia/verified-fetch [1.3.10](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-1.3.9...@helia/verified-fetch-1.3.10) (2024-04-12)
+
+
+### Documentation
+
+* fix readme link ([#51](https://github.com/ipfs/helia-verified-fetch/issues/51)) ([8a41c57](https://github.com/ipfs/helia-verified-fetch/commit/8a41c5701f800ddac25ae59d790836e5b1c48d93))
+
+## @helia/verified-fetch [1.3.9](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-1.3.8...@helia/verified-fetch-1.3.9) (2024-04-11)
+
+
+### Bug Fixes
+
+* identity CIDs use contentTypeParser ([#49](https://github.com/ipfs/helia-verified-fetch/issues/49)) ([3014498](https://github.com/ipfs/helia-verified-fetch/commit/30144981b5253b3269cbf186a24f1ef9dd04a452))
+
+## @helia/verified-fetch [1.3.8](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-1.3.7...@helia/verified-fetch-1.3.8) (2024-04-09)
+
+
+### Bug Fixes
+
+* pass url and body to badRequestResponse ([#44](https://github.com/ipfs/helia-verified-fetch/issues/44)) ([cc228e6](https://github.com/ipfs/helia-verified-fetch/commit/cc228e6fe74be0d340dc496191fe1dc06fd24486))
+
+
+### Dependencies
+
+* updating all deps ([#47](https://github.com/ipfs/helia-verified-fetch/issues/47)) ([6d0ffd8](https://github.com/ipfs/helia-verified-fetch/commit/6d0ffd837e15f1d5bb84a5b2c855d490301ac312))
+
+## @helia/verified-fetch [1.3.7](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-1.3.6...@helia/verified-fetch-1.3.7) (2024-04-08)
+
+
+### Bug Fixes
+
+* walking dag-cbor paths ([#39](https://github.com/ipfs/helia-verified-fetch/issues/39)) ([99668ce](https://github.com/ipfs/helia-verified-fetch/commit/99668cef249f223e143e359fb282c00d98b82f28))
+
+## @helia/verified-fetch [1.3.6](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-1.3.5...@helia/verified-fetch-1.3.6) (2024-04-08)
+
+
+### Bug Fixes
+
+* implicit accept header can be overridden by format query ([#36](https://github.com/ipfs/helia-verified-fetch/issues/36)) ([75c0b75](https://github.com/ipfs/helia-verified-fetch/commit/75c0b75f1b5d340b68063db0eab94c0228c261c4))
+
+## @helia/verified-fetch [1.3.5](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-1.3.4...@helia/verified-fetch-1.3.5) (2024-04-08)
+
+
+### Bug Fixes
+
+* remove redundant abort controller ([#41](https://github.com/ipfs/helia-verified-fetch/issues/41)) ([04b220d](https://github.com/ipfs/helia-verified-fetch/commit/04b220dab3ee9e617eb5ba920f4bb00112d1a861))
+
 ## @helia/verified-fetch [1.3.4](https://github.com/ipfs/helia-verified-fetch/compare/@helia/verified-fetch-1.3.3...@helia/verified-fetch-1.3.4) (2024-04-02)